Refactor move ConditionalSWR into its own file (#2618)
Moves `useConditionalSWR` into its own file.
