+++ frontend/src/hooks/api/getters/useEnterpriseSWR/useEnterpriseSWR.ts
@@ -1,21 +1,6 @@
-import useSWR, { BareFetcher, Key, SWRConfiguration, SWRResponse } from 'swr';
-import { useEffect } from 'react';
+import { BareFetcher, Key, SWRConfiguration } from 'swr';
+import { useConditionalSWR } from '../useConditionalSWR/useConditionalSWR';
 import useUiConfig from '../useUiConfig/useUiConfig';
-
-export const useConditionalSWR = <Data = any, Error = any, T = boolean>(
-    key: Key,
-    fetcher: BareFetcher<Data>,
-    condition: T,
-    options: SWRConfiguration = {}
-): SWRResponse<Data, Error> => {
-    const result = useSWR(key, fetcher, options);
-
-    useEffect(() => {
-        result.mutate();
-    }, [condition]);
-
-    return result;
-};
 
 export const useEnterpriseSWR = <Data = any, Error = any>(
     key: Key,
